Moved getScores inner class to its own outer class file
